Hopefully fixed formatting for ix_values.yaml
Hopefully this fixes the formatting and allows custom ix_values in the config file to update properly without getting in an update loop.

Be sure to remove quotes around values in your own config file for custom tag info, unless it is needed (For example if the tag line has a backslash, then be sure it has quotes around the value).

I accidentally had quotes around the values for Sonarr and it was partially causing an update loop. I also was applying quotes if it didn't have any which has now been fixed.